Combining more layers

+++ itsaren-simple/src/main/scala/com/hunorkovacs/itsaren/simple/User.scala
@@ -2,4 +2,4 @@ package com.hunorkovacs.itsaren.simple
 
 case class User(id: UserId, name: String)
 
-case class UserId(id: Int)
+case class UserId(id: Long)
